+++ CHANGELOG.rst
@@ -1,3 +1,16 @@
 
 Changelog
 =========
+
+--------------------
+[0.1.0] - 2021-06-09
+--------------------
+
+Added
+-----
+
+- requirements.txt file
+- README, CHANGELOG files 
+- functions.py with ini_binance and massive_ohlcv methods
+- main.py with first example
+